windows-client/requirements.txt
keyboard>=0.13.0
pyperclip>=1.8.2
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0 